+++ History.md
@@ -1,5 +1,5 @@
-unreleased
-==========
+0.1.0 / 2014-06-15
+==================
 
   * Add `deprecate.function(fn, message)`
   * Add `process.on('deprecation', fn)` emitter